Expand routes and add README.md

+++ backend/src/main/java/edu/lehigh/cse216/group4/backend/RequestIdea.java
@@ -5,5 +5,5 @@ public class RequestIdea {
     public String subject;
     public String content;
     public String attachment;
-    public Integer[] allowedRoles;
+    public Short[] allowedRoles;
 }
